separate webdriver launch from other logic
